feat(samples): Add and centralize sample modules under /samples
