+++ samples/app/proguard-rules.pro
