+++ samples/app/src/androidTest/java/com/harrytmthy/stitch/StitchVsDaggerBenchmark.kt
@@ -1,0 +1,132 @@
+/*
+ * Copyright 2025 Harry Timothy Tumalewa
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.harrytmthy.stitch
+
+import androidx.benchmark.ExperimentalBenchmarkConfigApi
+import androidx.benchmark.MicrobenchmarkConfig
+import androidx.benchmark.junit4.BenchmarkRule
+import androidx.benchmark.junit4.measureRepeated
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.harrytmthy.stitch.benchmark.BenchmarkComponent
+import com.harrytmthy.stitch.benchmark.DaggerBenchmarkComponent
+import com.harrytmthy.stitch.benchmark.DeepTarget
+import com.harrytmthy.stitch.benchmark.DeepTargetCold
+import com.harrytmthy.stitch.benchmark.ShallowTarget
+import com.harrytmthy.stitch.benchmark.ShallowTargetCold
+import com.harrytmthy.stitch.di.ViewModel
+import com.harrytmthy.stitch.generated.StitchDeepTargetColdInjector
+import com.harrytmthy.stitch.generated.StitchDeepTargetInjector
+import com.harrytmthy.stitch.generated.StitchShallowTargetColdInjector
+import com.harrytmthy.stitch.generated.StitchShallowTargetInjector
+import com.harrytmthy.stitch.generated.StitchViewModelInjector
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class StitchVsDaggerBenchmark {
+
+    @OptIn(ExperimentalBenchmarkConfigApi::class)
+    @get:Rule
+    val benchmarkRule = BenchmarkRule(
+        MicrobenchmarkConfig(
+            warmupCount = 20,
+            measurementCount = 100,
+        ),
+    )
+
+    private lateinit var daggerComponent: BenchmarkComponent
+
+    @Before
+    fun setUp() {
+        // Initialize Dagger component (generated by Dagger KSP)
+        daggerComponent = DaggerBenchmarkComponent.create()
+
+        // Stitch DI component + injectors are lazily initialized (on 1st access).
+        // Thus, this touch is needed for fairness.
+        StitchViewModelInjector.inject(ViewModel())
+    }
+
+    @Test
+    fun stitchShallowWarm() {
+        ShallowTarget().also(StitchShallowTargetInjector::inject) // Warmup
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { ShallowTarget() }
+                .also(StitchShallowTargetInjector::inject)
+        }
+    }
+
+    @Test
+    fun daggerShallowWarm() {
+        ShallowTarget().also(daggerComponent::inject) // Warmup
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { ShallowTarget() }
+                .also(daggerComponent::inject)
+        }
+    }
+
+    @Test
+    fun stitchDeepWarm() {
+        DeepTarget().also(StitchDeepTargetInjector::inject) // Warmup
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { DeepTarget() }
+                .also(StitchDeepTargetInjector::inject)
+        }
+    }
+
+    @Test
+    fun daggerDeepWarm() {
+        DeepTarget().also(daggerComponent::inject) // Warmup
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { DeepTarget() }
+                .also(daggerComponent::inject)
+        }
+    }
+
+    @Test
+    fun stitchShallowCold() {
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { ShallowTargetCold() }
+                .also(StitchShallowTargetColdInjector::inject)
+        }
+    }
+
+    @Test
+    fun daggerShallowCold() {
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { ShallowTargetCold() }
+                .also(daggerComponent::inject)
+        }
+    }
+
+    @Test
+    fun stitchDeepCold() {
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { DeepTargetCold() }
+                .also(StitchDeepTargetColdInjector::inject)
+        }
+    }
+
+    @Test
+    fun daggerDeepCold() {
+        benchmarkRule.measureRepeated {
+            runWithMeasurementDisabled { DeepTargetCold() }
+                .also(daggerComponent::inject)
+        }
+    }
+}
