+++ samples/app/src/androidTest/java/com/harrytmthy/stitch/StitchVsKoinBenchmark.kt
