+++ samples/app/src/main/java/com/harrytmthy/stitch/StitchSampleApp.kt
@@ -1,0 +1,7 @@
+package com.harrytmthy.stitch
+
+import android.app.Application
+import com.harrytmthy.stitch.annotations.StitchRoot
+
+@StitchRoot
+class StitchSampleApp : Application()
