+++ samples/app/src/main/java/com/harrytmthy/stitch/benchmark/BenchmarkClasses.kt
