+++ samples/app/src/main/java/com/harrytmthy/stitch/benchmark/BenchmarkComponent.kt
