+++ samples/app/src/main/java/com/harrytmthy/stitch/benchmark/BenchmarkModule.kt
