+++ samples/app/src/main/res/mipmap-mdpi/ic_launcher.webp
