+++ samples/app/src/main/res/values/strings.xml
