+++ samples/feature/home/build.gradle.kts
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2025 Harry Timothy Tumalewa
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+plugins {
+    alias(libs.plugins.stitch.android.library)
+    alias(libs.plugins.ksp)
+}
+
+android {
+    namespace = "com.harrytmthy.stitch.feature.home"
+
+    ksp {
+        arg("stitch.moduleName", "FeatureHome")
+    }
+}
+
+dependencies {
+    implementation(project(":core"))
+    implementation(project(":stitch"))
+    implementation(project(":stitch-annotations"))
+    implementation(libs.androidx.appcompat)
+    ksp(project(":stitch-compiler"))
+}
